requirements.txt
pandas==2.2.3
numpy==2.1.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
Flask==3.0.3
flask-cors==5.0.1
ruff==0.8.6
scipy==1.17.1
pytest==9.0.3
